requirements.txt
azure-functions==1.20.0
azure-storage-queue==12.10.0
psutil==5.9.8
requests==2.32.3
